implementation button delete in the listview

+++ app/src/main/java/com/app/nutritionapp/db/DBController.java
@@ -86,7 +86,7 @@ public class DBController {
         db.close();
     }
 
-    public void deleteRegister(int id){
+    public void deleteRegister(long id){
         String where = CreateDB.ID + "=" + id;
         db = banco.getReadableDatabase();
         db.delete(CreateDB.TABLE,where,null);
